extend our ui for a basic edit and add new, with some tabs so we can easily import DB entries for training data
