+++ admin_tabs/__init__.py
@@ -1,0 +1,1 @@
+# Admin tabs package
